requirements.txt
# ============================================================
# Insurance Claim AI - Python Dependencies
# ============================================================

# Backend Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# LLM Clients
openai==1.50.0
# anthropic==0.34.0  # Uncomment when Claude client is implemented

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# PDF Processing (Agent 0b - PDF Processing)
PyPDF2>=3.0.0
reportlab>=4.0.0
# pymupdf==1.24.0  # Alternative PDF library
# pdfplumber==0.11.0  # Alternative with better table extraction

# OCR & Image Processing (Agent 0a - Image Processing)
Pillow>=12.0.0
pytesseract>=0.3.13
# easyocr==1.7.1  # Optional - for advanced OCR

# ML & Deep Learning (Agent 1 - Requirements, Agent 2 - Credibility)
numpy>=2.0.0
scipy>=1.17.0
scikit-learn>=1.8.0
joblib>=1.5.0
torch>=2.11.0
transformers>=5.5.0

# NLP & Embeddings (Agent 1 - Document Similarity)
sentence-transformers>=5.3.0

# Vector Database (Agent 1 - Duplicate Detection)
faiss-cpu>=1.13.0

# Data Processing (Agent 2 - Credibility Model)
pandas>=2.3.0

# Token Counting
tiktoken>=0.7.0

# Validation
pydantic>=2.0.0

# Utilities
httpx==0.27.0
aiofiles==24.1.0

# Development
# pytest==8.3.0
# pytest-asyncio==0.24.0
# jupyter==1.0.0
# notebook==7.2.0
